% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonStandardStartStrength.R
\name{plot_distributions_experimentally_determined_start_strengths}
\alias{plot_distributions_experimentally_determined_start_strengths}
\title{Distributions of non-AUG start codon translation efficiency}
\usage{
plot_distributions_experimentally_determined_start_strengths()
}
\value{
visual summary of distribution of translation efficiencies of different start codons based on 8bp sequence context (ggplot)
}
\description{
Distributions of non-AUG start codon translation efficiency
}
\examples{
plot_distributions_experimentally_determined_start_strengths()
}
