% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonStandardStartStrength.R
\name{plot_start_start_strength_nonstandard}
\alias{plot_start_start_strength_nonstandard}
\title{Start Strength}
\usage{
plot_start_start_strength_nonstandard(input_sequence)
}
\arguments{
\item{input_sequence}{sequences of non-standard start sites to test. All input sequences must be 8bp long with start codon in pos 5bp-7bp, for example AAAA(UUG)A. No brackets should be included (character)}
}
\value{
Plot showing how Translation initiation efficiency in given contexts relates to the distribution of all possible sequence contexts
}
\description{
Use sequence context to identify the EXPERIMENTALLY determined tranlsation initiation start efficiency
Built using data from DOI: 10.1093/nar/gkx1114 . Please cite if you find this tool useful
}
\examples{
start_strength_nonstandard("TACTTTGG")
}
