% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startstrength-package.R
\docType{package}
\name{startstrength-package}
\alias{startstrength}
\alias{startstrength-package}
\title{startstrength: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Sam El-Kamand \email{selkamand@ccia.org.au} (\href{https://orcid.org/0000-0003-2270-8088}{ORCID})

}
\keyword{internal}
